Fixed null on protected values error
Ever so often we run into the problem of an property being sat by
default on a class when we don't actually want it set (for instance in
the case of an automatically updated value) and since it's set in the
class it can't really just be unset.

This fix lets everything that is set to null be ignored, if the field in
the database is not nullable.
This is concistant with php's own logic as it does not consider a
variable with the value of null as set.

This commit also adresses the rare instance of mom generating an invalid
SQL string if there is no values to actually add to the string.